got tournament page to show docId in url
