+++ app/tournaments/[id]/layout.js
